# grubhub-API-swagger-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CraveCrafter API",
    "description": "The API for the GrubHub part of the CraveCrafter API.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/simonpcdutton/CraveCrafter/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "paths": {
...
    "/diners/{diner_id}/favorites/restaurants": {
      "get": {
        "tags": [
          "Diners"
        ],
        "description": "Returns a list of restaurants that the currently logged in diner has favorited.",
        "operationId": "getFavoriteRestaurants",
        "parameters": [
          {
            "name": "diner_id",
            "in": "path",
            "description": "The ID of the diner whose favorites you want to retrieve.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListFavoriteRestaurantResponse"
                }
              }
            }
          }
        },
        "x-swagger-router-controller": "Diners"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FavoriteRestaurant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "An identifier for this favorite. This is not the restaurant ID"
          },
          "restaurant_id": {
            "type": "string",
            "description": "The unique ID of the favorited restaurant."
          },
          "create_time": {
            "type": "string",
            "description": "The time on which this restaurant was added to the diner's favorites.",
            "format": "timestamp"
          }
        },
        "description": "A restaurant that a diner has favorited along with the time that they favorited it."
      },
...
      "ListFavoriteRestaurantResponse": {
        "type": "object",
        "properties": {
          "favorite_restaurants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FavoriteRestaurant"
            }
          }
        },
        "description": "An array that contains a diner's favorited restaurants and when they favorited those restaurants."
      },
...
    }
  }
}